Compilation in Cygwin
git-svn-id: file:///home/roberto/Git/svn@10889 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0fa48925bb6268965128f273d13bbb013a6f86f8

+++ trunk/xmipp/libraries/reconstruction/tomo_align_tilt_series.cpp
@@ -2652,7 +2652,7 @@ void Alignment::computeGeometryDependentOfAxis()
 
         B=I-Raxis;
         B=B.transpose()*Ip*B;
-        Binvraxis+=prm->ni(i)*(B+B.transpose());
+        Binvraxis+=((double)prm->ni(i))*(B+B.transpose());
 
         B1i[i]=I-Raxis.transpose();
         B2i[i]=B1i[i]*Ip*Raxis;
